requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
multidict==6.0.4
mypy==1.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
ruff==0.1.5
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
yarl==1.9.2
